rework deploy script; move create2IfNotExists to utils
